requirements.txt
flask==3.0.3
firebase-admin>=6.0.0
python-dotenv>=1.0.0